refactor: `retriever` module
