+++ crates/learner/config/retrievers/iacr.toml
@@ -2,6 +2,7 @@ base_url          = "https://eprint.iacr.org"
 endpoint_template = "https://eprint.iacr.org/oai?verb=GetRecord&identifier=oai:eprint.iacr.org:{identifier}&metadataPrefix=oai_dc"
 name              = "iacr"
 pattern           = "(?:^|https?://eprint\\.iacr\\.org/)(\\d{4}/\\d+)(?:\\.pdf)?$"
+resource_type     = "paper"
 source            = "iacr"
 
 [response_format]
